Fix warnings due to signed/unsigned int comparison in CompTimer. (LP: #893998) Merged: lp:~smspillaz/compiz-core/fix-timer-warnings-893998 without any of its prereqs.
(bzr r2908.2.1)

+++ src/timer/src/timer.cpp
@@ -105,10 +105,10 @@ CompTimeoutSource::prepare (int &timeout)
 	while (it != TimeoutHandler::Default ()->timers ().end ())
 	{
 	    t = (*it);
-	    if (t->minLeft () >= timeout)
+	    if (t->minLeft () >= (unsigned int) timeout)
 		break;
-	    if (t->maxLeft () < timeout)
-		timeout = t->maxLeft ();
+	    if (t->maxLeft () < (unsigned int) timeout)
+		timeout = (int) t->maxLeft ();
 	    it++;
 	}
     }
